requirements.txt
numpy
pandas
openpyxl
xlrd
faker
scikit-learn
xgboost
joblib
PyPDF2
python-docx
shap
matplotlib
seaborn
plotly
pytest
fastapi
uvicorn[standard]
python-multipart
reportlab
python-jose[cryptography]
passlib[bcrypt]
httpx
